Improve Browserstack session cleanup
When cleaning up sessions in browserstack, the 'killCurrentSession'
function assumed that the current build is the first one of the running
builds returned by the API. This however may not always be true, eg. if
the same browserstack account is used for multiple projects, there may
be more than 1 builds running in parallel.

This commit improves the function by comparing the build name with the
one specified in desiredCapabilities (set in chimp config file). It's
however not mandatory to specify the build name, so if it's not
specified, the original assumption of selecting the first build still
stands.

In case of multiremote (multiple sessions), it's important to specify
the build name in desiredCapabilities, multiple sessions will not be
grouped under one build and hence only one of them will get killed. This
has been added as a note in the docstring.